requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
aiohttp=3.8.3=pypi_0
aiosignal=1.2.0=pypi_0
anyio=3.6.2=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.2=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2022.9.24=py37h06a4308_0
cffi=1.15.1=pypi_0
chardet=4.0.0=pypi_0
charset-normalizer=2.1.1=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
d2l=0.17.6=pypi_0
debugpy=1.6.6=pypi_0
decorator=5.1.1=pypi_0
deepspeed=0.7.4=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.4=pypi_0
fair-esm=2.0.0=pypi_0
fastjsonschema=2.16.3=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.1=pypi_0
fsspec=2022.10.0=pypi_0
giflib=5.2.1=h7b6447c_0
google-auth=2.13.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.50.0=pypi_0
hjson=3.1.0=pypi_0
idna=2.10=pypi_0
importlib-metadata=5.0.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.16.2=pypi_0
ipython=7.34.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.4=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.4.9=pypi_0
jupyter-console=6.6.1=pypi_0
jupyter-core=4.12.0=pypi_0
jupyter-server=1.23.6=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.5=pypi_0
kiwisolver=1.4.4=pypi_0
lcms2=2.12=h3be6417_0
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.5=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
multidict=6.0.2=pypi_0
nbclassic=0.5.2=pypi_0
nbclient=0.7.2=pypi_0
nbconvert=7.2.9=pypi_0
nbformat=5.7.3=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.6=pypi_0
ninja=1.10.2.4=pypi_0
notebook=6.5.2=pypi_0
notebook-shim=0.2.2=pypi_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
oauthlib=3.2.2=pypi_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.2.4=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=py37hace64e9_1
pip=22.2.2=py37h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.37=pypi_0
protobuf=3.19.6=pypi_0
psutil=5.9.3=pypi_0
ptyprocess=0.7.0=pypi_0
py-cpuinfo=8.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pydantic=1.10.2=pypi_0
pydeprecate=0.3.2=pypi_0
pygments=2.14.0=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.7.2=h0371630_0
python-dateutil=2.8.2=pypi_0
pytorch=1.12.1=py3.7_cuda11.3_cudnn8.3.2_0
pytorch-lightning=1.7.7=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.5=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.0=pypi_0
qtconsole=5.4.0=pypi_0
qtpy=2.3.0=pypi_0
readline=7.0=h7b6447c_5
requests=2.25.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
send2trash=1.8.0=pypi_0
setuptools=63.4.1=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4=pypi_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.12.1=py37_cu113
torchmetrics=0.10.0=pypi_0
torchvision=0.2.2=py_3
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.9.0=pypi_0
typing_extensions=4.3.0=py37h06a4308_0
urllib3=1.26.12=pypi_0
wcwidth=0.2.6=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.5=pypi_0
xz=5.2.6=h5eee18b_0
yarl=1.8.1=pypi_0
zipp=3.9.0=pypi_0
zlib=1.2.12=h5eee18b_3
zstd=1.5.2=ha4553b6_0
